% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tAUC.R
\name{data_crossing}
\alias{data_crossing}
\title{Generate a pseudo data}
\usage{
data_crossing(vt, vc, vm, vxs)
}
\arguments{
\item{vt}{Vector of time}

\item{vc}{Vector of censoring time}

\item{vm}{Vector of biomarker}

\item{vxs}{Matrix of biomarkers}
}
\value{
tableij
}
\description{
Generate a pseudo data
}
