% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tAUC.R
\name{se.cal}
\alias{se.cal}
\title{Standard Error estimation}
\usage{
se.cal(beta.hat, V, t0, nf = 7)
}
\arguments{
\item{beta.hat}{coefficients}

\item{V}{covariance matrix}

\item{t0}{time interest}

\item{nf}{number of elements of fractional polynomial}
}
\value{
se
}
\description{
Standard Error estimation
}
